pyproject.toml

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "BotScanner"
version = "0.3.0"
description = "Automation framework for scanning, vault, and firewall routines"
authors = [{ name = "Leon McClatchey" }]
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.2",
    "aiosignal>=1.4.0",
    "ansible>=12.2.0",
    "ansible-core>=2.19.4",
    "ansible-vault>=4.1.0",
    "attrs>=25.4.0",
    "certifi>=2025.11.12",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "cryptography>=46.0.3",
    "defusedxml>=0.7.1",
    "distro>=1.9.0",
    "dnspython>=2.8.0",
    "frozenlist>=1.8.0",
    "geoip2>=5.2.0",
    "idna>=3.11",
    "iniconfig>=2.3.0",
    "ipwhois>=1.3.0",
    "Jinja2>=3.1.6",
    "MarkupSafe>=3.0.3",
    "maxminddb>=3.0.0",
    "multidict>=6.7.0",
    "packaging>=25.0",
    "pluggy>=1.6.0",
    "propcache>=0.4.1",
    "psutil>=7.1.3",
    "pycparser>=2.23",
    "Pygments>=2.19.2",
    "pytest>=9.0.1",
    "PyYAML>=6.0.3",
    "requests>=2.32.5",
    "resolvelib>=1.2.1",
    "typing_extensions>=4.15.0",
    "urllib3>=2.6.2",
    "yarl>=1.22.0"
]
[tool.setuptools.packages.find]
include = ["BotScanner*"]
exclude = ["data*", "Logs*", "archive*"]
